+++ sim/bevy_panorbit_camera/examples/multiple_windows.rs
@@ -1,8 +1,7 @@
 //! Demonstrates usage with multiple windows
 
-use bevy::prelude::*;
-use bevy::render::camera::RenderTarget;
 use bevy::window::WindowRef;
+use bevy::{camera::RenderTarget, prelude::*};
 use bevy_panorbit_camera::{PanOrbitCamera, PanOrbitCameraPlugin};
 
 fn main() {
@@ -32,7 +31,7 @@ fn setup(
     // Light
     commands.spawn((
         PointLight {
-            shadows_enabled: true,
+            shadow_maps_enabled: true,
             ..default()
         },
         Transform::from_xyz(4.0, 8.0, 4.0),
@@ -54,10 +53,8 @@ fn setup(
     // second window camera
     commands.spawn((
         Transform::from_translation(Vec3::new(5.0, 1.5, 7.0)),
-        Camera {
-            target: RenderTarget::Window(WindowRef::Entity(second_window)),
-            ..default()
-        },
+        Camera { ..default() },
+        RenderTarget::Window(WindowRef::Entity(second_window)),
         PanOrbitCamera::default(),
     ));
 }